VA branch issues: counters + save log
- Counters:
It was found that the same exact counter was used for all items that were
converted to use globals rather than local variables.  That counter was
@eval@("status","loaded") & @eval@("status","errors"). This caused us to get
bad numbers for each item as they were all effectively combined together.
- Save Log:
Immunizations, Allergies, and Problems were load log was not saved.
Investigation eventually revealed that the save actually did indeed happen, but
it was killed after the save as the save was done in two spots, each with a
kill before it; and the last kill has no data to save. We removed the incorrect
spot.