Fix lab import from Synthea
Due to the latest changes in Synthea, labs now don't import anymore and
cause an infinite loop. The cause of this was decimal points on labs,
which cause VistA to reject them; and the code is not well written... a
rejection means an infinite try and try again. So the temporary fix
(until we can get better code) is to round off labs to integers, which
is bad for potassium levels, but good enough right now.